debug

+++ src/osyris/io/amr.py
@@ -148,8 +148,8 @@ class AmrReader(Reader):
         self.offsets['n'] += 1 + 2 * ndim
 
     def read_variables(self, ncache, ind, ilevel, cpuid, info):
-        begin = ind * ncache
-        end = (ind + 1) * ncache
+        begin = int(ind * ncache)
+        end = int((ind + 1) * ncache)
         self.son[begin:end] = utils.read_binary_data(fmt="{}i".format(ncache),
                                                      content=self.bytes,
                                                      offsets=self.offsets)
